+++ CHANGELOG.md
@@ -8,6 +8,8 @@ This project uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ### Added
 
 - Added implementation of `From<Vec<T>>` trait.
+- Added `GenerationalVector::<T>::new_from_iter()` to construct from
+  arbitrary `IntoIterator<Item = T>` types.
 - Added `IntoIter` implementations for `GenerationalVector<T>`, `&GenerationalVector<T>` and `&mut GenerationalVector<T>`.
 
 ## 0.1.0 - 2022-04-15